refactor(alerts): split notification formatting
Move accessible alert notification message formatting and settings debug summaries into a focused helper module while preserving alert_notification_system imports.

Tested: pytest tests/test_alert_notification_system.py tests/test_alert_manager.py tests/test_alert_lifecycle.py tests/test_notification_test.py tests/test_app_notification_activation.py; ruff check src tests scripts; pyright.

Confidence: high
